windows/requirements.txt
# Auto-Stop Media - Windows Service Dependencies
# Python 3.9+ required

# WebSocket server
websockets>=12.0

# Windows Runtime - prebuilt wheels (no Visual Studio required)
winrt-runtime>=3.0.0
winrt-Windows.Foundation>=3.0.0
winrt-Windows.Foundation.Collections>=3.0.0
winrt-Windows.Media.Control>=3.0.0
winrt-Windows.Storage.Streams>=3.0.0

# System tray icon
pystray>=0.19.0
Pillow>=10.0.0

# Audio session detection (for Spotify/Spicetify fallback)
pycaw>=20230407
psutil>=5.9.0

